Add constructor and implement getBalance

+++ src/main/java/com/priceminister/account/Account.java
@@ -1,5 +1,7 @@
 package com.priceminister.account;
 
+import java.math.BigDecimal;
+
 /**
  * This class represents a simple account.
  * It doesn't handle different currencies, all money is supposed to be of standard currency EUR.
@@ -10,7 +12,7 @@ public interface Account {
      * Adds money to this account.
      * @param addedAmount - the money to add
      */
-    public void add(Double addedAmount);
+    public void add(BigDecimal addedAmount);
     
     /**
      * Withdraws money from the account.
@@ -19,11 +21,11 @@ public interface Account {
      * @return the remaining account balance
      * @throws IllegalBalanceException if the withdrawal leaves the account with a forbidden balance
      */
-    public Double withdrawAndReportBalance(Double withdrawnAmount, AccountRule rule) throws IllegalBalanceException;
+    public BigDecimal withdrawAndReportBalance(BigDecimal withdrawnAmount, AccountRule rule) throws IllegalBalanceException;
     
     /**
      * Gets the current account balance.
      * @return the account's balance
      */
-    public Double getBalance();
+    public BigDecimal getBalance();
 }
